+++ src/Network/Wai/Middleware/Delegate.hs
@@ -1,4 +1,7 @@
+{-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE RankNTypes #-}
 
 {- |
 Copyright   : (c) 2018-2021 Tim Emiola
@@ -19,76 +22,84 @@ Provides 3 combinators that create middleware along with supporting data types.
 * 'simpleProxy': is a simple reverse proxy, based on proxyApp of http-proxy by
   Erik de Castro Lopo/Michael Snoyman
 -}
-module Network.Wai.Middleware.Delegate (
-  -- * Middleware
-  delegateTo,
-  delegateToProxy,
-  simpleProxy,
+module Network.Wai.Middleware.Delegate
+  ( -- * Middleware
+    delegateTo
+  , delegateToProxy
+  , simpleProxy
 
-  -- * Configuration
-  ProxySettings (..),
+    -- * Configuration
+  , ProxySettings (..)
 
-  -- * Aliases
-  RequestPredicate,
-) where
+    -- * Aliases
+  , RequestPredicate
+  )
+where
 
 import Blaze.ByteString.Builder (fromByteString)
 import Control.Concurrent.Async (race_)
-import Control.Exception (
-  SomeException,
-  handle,
-  toException,
- )
+import Control.Exception
+  ( SomeException
+  , handle
+  , toException
+  )
+import Control.Monad (unless)
 import Control.Monad.IO.Class (MonadIO, liftIO)
+import Data.ByteString (ByteString)
 import qualified Data.ByteString as BS
 import qualified Data.ByteString.Char8 as C8
 import qualified Data.ByteString.Lazy.Char8 as LC8
 import Data.CaseInsensitive (mk)
-import Data.Conduit (
-  ConduitT,
-  Flush (..),
-  Void,
-  mapOutput,
-  runConduit,
-  yield,
-  (.|),
- )
+import Data.Conduit
+  ( ConduitM
+  , ConduitT
+  , Flush (..)
+  , Void
+  , await
+  , mapOutput
+  , runConduit
+  , yield
+  , ($$+)
+  , ($$++)
+  , (.|)
+  )
 import Data.Conduit.Network (appSink, appSource)
 import Data.Default (Default (..))
+import Data.IORef (newIORef, readIORef, writeIORef)
+import Data.Int (Int64)
 import Data.Monoid ((<>))
-import Data.Streaming.Network (
-  ClientSettings,
-  clientSettingsTCP,
-  runTCPClient,
- )
+import Data.Streaming.Network
+  ( ClientSettings
+  , clientSettingsTCP
+  , runTCPClient
+  )
 import Data.String (IsString)
-import Network.HTTP.Client (
-  Manager,
-  Request (..),
-  Response (..),
-  parseRequest,
-  withResponse,
- )
-import Network.HTTP.Client.Conduit (bodyReaderSource)
-import Network.HTTP.Conduit (
-  requestBodySourceChunkedIO,
-  requestBodySourceIO,
- )
-import Network.HTTP.Simple (Header)
-import Network.HTTP.Types (
-  HeaderName,
-  hContentType,
-  internalServerError500,
-  status304,
-  status500,
- )
+import Network.HTTP.Client
+  ( BodyReader
+  , GivesPopper
+  , Manager
+  , Request (..)
+  , RequestBody (..)
+  , Response (..)
+  , brRead
+  , parseRequest
+  , withResponse
+  )
+import Network.HTTP.Types
+  ( Header
+  , HeaderName
+  , hContentType
+  , internalServerError500
+  , status304
+  , status500
+  )
 import Network.HTTP.Types.Header (hHost)
 import qualified Network.Wai as Wai
-import Network.Wai.Conduit (
-  responseRawSource,
-  responseSource,
-  sourceRequestBody,
- )
+import Network.Wai.Conduit
+  ( responseRawSource
+  , responseSource
+  , sourceRequestBody
+  )
 
 
 {- | Type alias for a function that determines if a request should be handled by
@@ -115,14 +126,14 @@ delegateToProxy settings mgr = delegateTo (simpleProxy settings mgr)
 
 -- | Settings that configure the proxy endpoint.
 data ProxySettings = ProxySettings
-  { -- | What to do with exceptions thrown by either the application or server.
-    proxyOnException :: SomeException -> Wai.Response
-  , -- | Timeout value in seconds. Default value: 30
-    proxyTimeout :: Int
-  , -- | The host being proxied
-    proxyHost :: BS.ByteString
-  , -- | The number of redirects to follow. 0 means none, which is the default.
-    proxyRedirectCount :: Int
+  { proxyOnException :: SomeException -> Wai.Response
+  -- ^ What to do with exceptions thrown by either the application or server.
+  , proxyTimeout :: Int
+  -- ^ Timeout value in seconds. Default value: 30
+  , proxyHost :: BS.ByteString
+  -- ^ The host being proxied
+  , proxyRedirectCount :: Int
+  -- ^ The number of redirects to follow. 0 means none, which is the default.
   }
 
 
@@ -154,46 +165,46 @@ simpleProxy settings manager req respond
   -- we may connect requests to secure sites, when we do, we will not have
   -- seen their URI properly
   | Wai.requestMethod req == "CONNECT" =
-    respond $
-      responseRawSource
-        (handleConnect req)
-        (Wai.responseLBS status500 [("Content-Type", "text/plain")] "method CONNECT is not supported")
+      respond $
+        responseRawSource
+          (handleConnect req)
+          (Wai.responseLBS status500 [("Content-Type", "text/plain")] "method CONNECT is not supported")
   | otherwise = do
-    let scheme
-          | Wai.isSecure req = "https"
-          | otherwise = "http"
-        rawUrl = Wai.rawPathInfo req <> Wai.rawQueryString req
-        effectiveUrl = scheme ++ "://" ++ C8.unpack (proxyHost settings) ++ C8.unpack rawUrl
-        newHost = proxyHost settings
-        addHostHeader = (:) (hHost, newHost)
+      let scheme
+            | Wai.isSecure req = "https"
+            | otherwise = "http"
+          rawUrl = Wai.rawPathInfo req <> Wai.rawQueryString req
+          effectiveUrl = scheme ++ "://" ++ C8.unpack (proxyHost settings) ++ C8.unpack rawUrl
+          newHost = proxyHost settings
+          addHostHeader = (:) (hHost, newHost)
 
-    proxyReq' <- parseRequest effectiveUrl
-    let onException :: SomeException -> Wai.Response
-        onException = proxyOnException settings . toException
+      proxyReq' <- parseRequest effectiveUrl
+      let onException :: SomeException -> Wai.Response
+          onException = proxyOnException settings . toException
 
-        proxyReq =
-          proxyReq'
-            { method = Wai.requestMethod req
-            , requestHeaders = addHostHeader $ filter dropUpstreamHeaders $ Wai.requestHeaders req
-            , -- always pass redirects back to the client.
-              redirectCount = proxyRedirectCount settings
-            , requestBody =
-                case Wai.requestBodyLength req of
-                  Wai.ChunkedBody ->
-                    requestBodySourceChunkedIO (sourceRequestBody req)
-                  Wai.KnownLength l ->
-                    requestBodySourceIO (fromIntegral l) (sourceRequestBody req)
-            , -- don't modify the response to ensure consistency with the response headers
-              decompress = const False
-            , host = newHost
-            }
+          proxyReq =
+            proxyReq'
+              { method = Wai.requestMethod req
+              , requestHeaders = addHostHeader $ filter dropUpstreamHeaders $ Wai.requestHeaders req
+              , -- always pass redirects back to the client.
+                redirectCount = proxyRedirectCount settings
+              , requestBody =
+                  case Wai.requestBodyLength req of
+                    Wai.ChunkedBody ->
+                      requestBodySourceChunked (sourceRequestBody req)
+                    Wai.KnownLength l ->
+                      requestBodySource (fromIntegral l) (sourceRequestBody req)
+              , -- don't modify the response to ensure consistency with the response headers
+                decompress = const False
+              , host = newHost
+              }
 
-        respondUpstream = withResponse proxyReq manager $ \res -> do
-          let body = mapOutput (Chunk . fromByteString) . bodyReaderSource $ responseBody res
-              headers = (mk "X-Via-Proxy", "yes") : responseHeaders res
-          respond $ responseSource (responseStatus res) headers body
+          respondUpstream = withResponse proxyReq manager $ \res -> do
+            let body = mapOutput (Chunk . fromByteString) . bodyReaderSource $ responseBody res
+                headers = (mk "X-Via-Proxy", "yes") : responseHeaders res
+            respond $ responseSource (responseStatus res) headers body
 
-    handle (respond . onException) respondUpstream
+      handle (respond . onException) respondUpstream
 
 
 handleConnect ::
@@ -230,3 +241,45 @@ dropUpstreamHeaders (k, _) = k `notElem` preservedHeaders
 
 preservedHeaders :: [HeaderName]
 preservedHeaders = ["content-encoding", "content-length", "host"]
+
+
+type Source' = ConduitT () ByteString IO ()
+
+
+srcToPopperIO :: Source' -> GivesPopper ()
+srcToPopperIO src f = do
+  (rsrc0, ()) <- src $$+ return ()
+  irsrc <- newIORef rsrc0
+  let popper :: IO ByteString
+      popper = do
+        rsrc <- readIORef irsrc
+        (rsrc', mres) <- rsrc $$++ await
+        writeIORef irsrc rsrc'
+        case mres of
+          Nothing -> return BS.empty
+          Just bs
+            | BS.null bs -> popper
+            | otherwise -> return bs
+  f popper
+
+
+requestBodySource :: Int64 -> Source' -> RequestBody
+requestBodySource size = RequestBodyStream size . srcToPopperIO
+
+
+requestBodySourceChunked :: Source' -> RequestBody
+requestBodySourceChunked = RequestBodyStreamChunked . srcToPopperIO
+
+
+bodyReaderSource ::
+  (MonadIO m) =>
+  BodyReader ->
+  ConduitT i ByteString m ()
+bodyReaderSource br =
+  loop
+  where
+    loop = do
+      bs <- liftIO $ brRead br
+      unless (BS.null bs) $ do
+        yield bs
+        loop