+++ ChangeLog.md
@@ -2,6 +2,12 @@
 
 `wai-middleware-delegate` uses [PVP Versioning][1].
 
+## 0.1.4.0 -- 2023-17-11
+
+* Relax the upper bounds on bytestring to allow bytestring 0.12.1
+* Remove dependencies on http-conduit and connection
+* Switch to running the integration tests using http-bin in docker
+
 ## 0.1.3.1 -- 2022-08-09
 
 * Relax the version bounds so that pre 2.0 versions of text are still supported